requirements.txt
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
scikit-learn==1.4.2
scikit-surprise==1.1.4
matplotlib==3.8.4
seaborn==0.13.2
requests==2.32.3
tqdm==4.66.4